# translate netlib l2norm algorithm to asm and remove branches