--- a/internal/asm/f64/stubs_amd64.go
+++ b/internal/asm/f64/stubs_amd64.go
@@ -170,3 +170,29 @@ func ScalIncTo(dst []float64, incDst uintptr, alpha float64, x []float64, n, inc
 //      sum += x[i]
 //  }
 func Sum(x []float64) float64
+
+// L2NormUnitary returns the L2-norm of x.
+//   var scale float64
+//   sumSquares := 1.0
+//   for _, v := range x {
+//   	if v == 0 {
+//   		continue
+//   	}
+//   	absxi := math.Abs(v)
+//   	if math.IsNaN(absxi) {
+//   		return math.NaN()
+//   	}
+//   	if scale < absxi {
+//   		s := scale / absxi
+//   		sumSquares = 1 + sumSquares*s*s
+//   		scale = absxi
+//   	} else {
+//   		s := absxi / scale
+//   		sumSquares += s * s
+//   	}
+// 	  	if math.IsInf(scale, 1) {
+// 		  	return math.Inf(1)
+// 	  	}
+//   }
+//   return scale * math.Sqrt(sumSquares)
+func L2NormUnitary(x []float64) (sum float64)