Netlib algorithm reduces overflow while calculating the l2norm of a
vector.

Translated to asm while reducing branches in NaN and Inf checks.
Overflow protection is equivalent to the Netlib standard implementation.

Builds upon #1134 to further speed up L2Norm operations.

<details>
<summary>Speedup ~-33%</summary>

```
name                                 old time/op    new time/op    delta
L2NormNetlib/L2NormUnitary-1-8         5.08ns ± 2%    5.20ns ± 3%   +2.36%  (p=0.019 n=10+10)
L2NormNetlib/L2NormUnitary-3-8         9.00ns ± 2%    8.64ns ± 3%   -3.95%  (p=0.000 n=10+10)
L2NormNetlib/L2NormUnitary-10-8        27.6ns ± 4%    22.1ns ± 1%  -19.84%  (p=0.000 n=10+9)
L2NormNetlib/L2NormUnitary-30-8        89.5ns ± 6%    61.8ns ± 1%  -31.00%  (p=0.000 n=10+10)
L2NormNetlib/L2NormUnitary-100-8        314ns ± 1%     213ns ± 1%  -32.06%  (p=0.000 n=9+9)
L2NormNetlib/L2NormUnitary-300-8        977ns ± 3%     650ns ± 2%  -33.47%  (p=0.000 n=10+10)
L2NormNetlib/L2NormUnitary-1000-8      3.28µs ± 2%    2.17µs ± 2%  -33.83%  (p=0.000 n=10+10)
L2NormNetlib/L2NormUnitary-3000-8      9.89µs ± 1%    6.47µs ± 0%  -34.58%  (p=0.000 n=9+9)
L2NormNetlib/L2NormUnitary-10000-8     33.0µs ± 1%    21.6µs ± 0%  -34.62%  (p=0.000 n=8+8)
L2NormNetlib/L2NormUnitary-30000-8     97.9µs ± 1%    64.9µs ± 1%  -33.71%  (p=0.000 n=10+10)
L2NormNetlib/L2NormUnitary-100000-8     323µs ± 0%     216µs ± 0%  -33.27%  (p=0.000 n=9+8)
```
</details> 

Please take a look.

edit: 

Though this doesn't directly change any floats code, updates via #1134:
<details>
<summary>Floats</summary>

```
name           old time/op  new time/op  delta
Norm2Small-8   27.6ns ± 7%  21.1ns ± 8%  -23.66%  (p=0.000 n=10+10)
Norm2Medium-8  2.12µs ± 3%  1.24µs ± 2%  -41.39%  (p=0.000 n=10+9)
Norm2Large-8    205µs ± 2%   122µs ± 2%  -40.27%  (p=0.000 n=10+10)
Norm2Huge-8    21.1ms ± 2%  13.0ms ± 3%  -38.74%  (p=0.000 n=10+10)
```
</details>